+++ tests/test_insert.py
@@ -10,43 +10,37 @@ class TestInsert:
     vectors = [[random.random() for _ in range(dim)] for _ in range(10000)]
 
     def test_insert_normal(self, gcon, gcollection):
-        status, ids = gcon.insert(gcollection, self.vectors)
-
-        assert status.OK()
-        assert len(ids) == len(self.vectors)
+        try:
+            ids = gcon.insert(gcollection, self.vectors)
+            assert len(ids) == len(self.vectors)
+        except Exception:
+            pytest.fail("Fail raise")
 
     def test_insert_with_ids(self, gcon, gcollection):
         ids = [i for i in range(10000)]
-        status, ids_ = gcon.insert(gcollection, self.vectors, ids)
-
-        assert status.OK()
-        assert len(ids_) == len(self.vectors)
-        assert ids == ids_
+        try:
+            ids_ = gcon.insert(gcollection, self.vectors, ids)
+            assert len(ids_) == len(self.vectors)
+            assert ids == ids_
+        except Exception:
+            pytest.fail("Fail raise")
 
     def test_insert_with_partition(self, gcon, gcollection):
-        status = gcon.create_partition(gcollection, "tag01")
-        assert status.OK()
-
-        status, ids = gcon.insert(gcollection, self.vectors, partition_tag="tag01")
-
-        assert status.OK()
+        gcon.create_partition(gcollection, "tag01")
+        ids = gcon.insert(gcollection, self.vectors, partition_tag="tag01")
         assert len(ids) == len(self.vectors)
 
     def test_insert_async(self, gcon, gcollection):
         future = gcon.insert(gcollection, self.vectors, _async=True)
-
-        status, ids = future.result()
-        assert status.OK()
+        ids = future.result()
         assert len(ids) == len(self.vectors)
 
     def test_insert_async_callback(self, gcon, gcollection):
-        def cb(status, ids):
-            assert status.OK()
+        def cb(ids):
             assert len(ids) == len(self.vectors)
 
         future = gcon.insert(gcollection, self.vectors, _async=True, _callback=cb)
-        status, _ = future.result()
-        assert status.OK()
+        future.result()
         future.done()
 
     @pytest.mark.parametrize("vectors", [[], None, "", 12344, [1111], [[]], [[], [1.0, 2.0]]])